+++ source/fixes/ProjectIGI1WidescreenFix.cpp
@@ -80,7 +80,7 @@ void HandleFOVInput(double &newCustomFOV)
         replace(input.begin(), input.end(), ',', '.');
 
         // Parses the string to a double
-        newCustomFOV = stof(input);
+        newCustomFOV = stod(input);
 
         if (cin.fail())
         {
