+++ source/fixes/Spiderman2TheGameFOVFix.cpp
@@ -83,7 +83,7 @@ void HandleFOVInput(double &newCustomFOV)
         replace(input.begin(), input.end(), ',', '.');
 
         // Parses the string to a double
-        newCustomFOV = stof(input);
+        newCustomFOV = stod(input);
 
         if (cin.fail())
         {
